✨Feat:  orders  protected

+++ backend/controllers/ordersControllers.js
@@ -3,7 +3,7 @@ const Order = require('../models/ordersModel')
 
 const getOrders = asyncHandler(async (req, res) =>  {
 
-    const orders = await Order.find()
+    const orders = await Order.find({user: req.user.id})
     res.status(200).json(orders)
 })
 
@@ -14,7 +14,8 @@ const setOrder = asyncHandler(async (req, res) =>  {
         
     }
     const order = await Order.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(201).json(order)
@@ -28,9 +29,18 @@ const updateOrder =  asyncHandler(async(req, res) =>  {
         throw new Error('The order was not found')
     }
 
-    const updatedOrder =  await Order.findByIdAndUpdate(req.params.id, req.body, {new: true})
+    //Check that the order belongs to the user's login
 
-    res.status(200).json(updatedOrder)
+    if(order.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error('unauthorized user')
+    }else{
+        const updatedOrder =  await Order.findByIdAndUpdate(req.params.id, req.body, {new: true})
+
+        res.status(200).json(updatedOrder)
+    }
+
+    
 })
 
 const deleteOrder = asyncHandler(async(req, res) =>  {
@@ -41,12 +51,22 @@ const deleteOrder = asyncHandler(async(req, res) =>  {
         res.status(404)
         throw new Error('The order was not found')
     }
+
+    if(order.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error('unauthorized user')
+    }else{
+        order.deleteOne()
+
+    res.status(200).json({id: order._id})
+       
+    }
     
     //const deletedOrder = await Order.findByIdAndDelete(req.params.id)
 
-    order.deleteOne()
+    // order.deleteOne()
 
-    res.status(200).json({id: order._id})
+    // res.status(200).json({id: order._id})
 })
 
 module.exports = {
